requirements/requirements-prod.txt
-r requirements-base.txt

# whitenoise>=5.2,<6
django-on-heroku>=1.1.2,<2
psycopg2-binary>=2.8.6,<3.0.0
gunicorn>=20.0.4,<21.0.0